+++ docs/source/changelog.rst
@@ -32,6 +32,8 @@ ChangeLog
 * Make the Key Usage extension in :command:`manage.py init_ca` configurable.
 * Add support for adding the Inhibit anyPolicy and Policy Constraints extensions when using
   :command:`manage.py init_ca`.
+* Add support for adding the OCSP No Check extension in when issuing certificates with :command:`manage.py
+  sign_cert` or :command:`manage.py resign_cert`.
 * Add support for multiple OCSP responder and CA Issuer entries when creating a certificate authority.
 * Add typehints when installing as wheel.